Add interface to configure client later
For webapps it is quite common to want to use a global XMPP client instance,
but not have credentials for it until a later point in time. Currently this
is not supported via @xmpp/client: it assumes you have the complete
configuration at construction time.

You can use @xmpp/client-core to handle this, but that would force you to
duplicate all the logic in @xmpp/client every time.

As an alternative we split the client creation in two parts: creating a client,
and configuring its options.