fix(tls): Verify certificates
Only direct TLS is affected (xmpps: protocol or via xmpps SRVs).

STARTTLS is not affected.
WebSocket is not affected.

Node.js tls API was used incorrectly, as a consequence the tls certificates were not verified by the client when connecting using direct tls.
It was equivalent to setting NODE_TLS_REJECT_UNAUTHORIZED=0